Fix yarn test-contracts command
Due to the fact that we set some configuration in `truffle-config`
truffle is looking for a running ethereum client to run the tests
against. I changed the `yarn test-contracts` in a way that it uses its
internal `truffle develop` to run the tests. So I would suggest when
running tests locally with a running ethereum client run
`truffle test`, which uses the specified config in truffle-config.
When running tests via `yarn test-contracts` it will run the tests in
clean room env.

Fixes
https://github.com/cp-ss2019/decentralized-energy-trading/issues/16.